+++ CONY-BE/payment/src/main/java/com/cony/payment/domain/sale/dto/SaleListResponseDto.java
@@ -23,8 +23,8 @@ public class SaleListResponseDto {
     private Long gifticonId;
 
     // 가격 정보
-    private Long originalPrice;      // 정가
-    private Long salePrice;          // 판매가
+    private Integer originalPrice;      // 정가
+    private Integer salePrice;          // 판매가
     private Double discountRate;     // 할인율
 
     // 기프티콘 정보 (Manage 서버에서 조회)
